fix: compare plugin version with equity
Previous bug caused issues when requiring exact minimum version of
plugin specified by SemVer X.Y.Z, as Z part was not compared for equity,
but was only checked if the version was lower. This led to situations
when requiring (and having activated) plugin in 1.1.1 version actually
required plugin in version 1.1.2 to work correctly.

Signed-off-by: Bart Jaskulski <bjaskulski@protonmail.com>